requirements.txt
streamlit==1.32.0
google-generativeai==0.8.3
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0